FastAPI/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
protobuf>=5.26.1
groq

# Run dev server:
# uvicorn app.main:app --reload --host 0.0.0.0 --port 8000
# venv\Scripts\activate
